// src/generated/wotlk/ui.ts
export interface DatabaseFilters {
    /**
     * @generated from protobuf field: repeated proto.ArmorType armor_types = 1
     */
    armorTypes: ArmorType[];
    /**
     * @generated from protobuf field: repeated proto.WeaponType weapon_types = 2
     */
    weaponTypes: WeaponType[];
    /**
     * @generated from protobuf field: repeated proto.RangedWeaponType ranged_weapon_types = 16
     */
    rangedWeaponTypes: RangedWeaponType[];
    /**
     * @generated from protobuf field: repeated proto.SourceFilterOption sources = 17
     */
    sources: SourceFilterOption[];
    /**
     * @generated from protobuf field: repeated proto.RaidFilterOption raids = 18
     */
    raids: RaidFilterOption[];
    /**
     * @generated from protobuf field: proto.UIItem.FactionRestriction faction_restriction = 19
     */
    factionRestriction: UIItem_FactionRestriction;
    /**
     * @generated from protobuf field: double min_mh_weapon_speed = 4
     */
    minMhWeaponSpeed: number;
    /**
     * @generated from protobuf field: double max_mh_weapon_speed = 5
     */
    maxMhWeaponSpeed: number;
    /**
     * @generated from protobuf field: double min_oh_weapon_speed = 9
     */
    minOhWeaponSpeed: number;
    /**
     * @generated from protobuf field: double max_oh_weapon_speed = 10
     */
    maxOhWeaponSpeed: number;
    /**
     * @generated from protobuf field: double min_ranged_weapon_speed = 14
     */
    minRangedWeaponSpeed: number;
    /**
     * @generated from protobuf field: double max_ranged_weapon_speed = 15
     */
    maxRangedWeaponSpeed: number;
    /**
     * @generated from protobuf field: bool one_handed_weapons = 6
     */
    oneHandedWeapons: boolean;
    /**
     * @generated from protobuf field: bool two_handed_weapons = 7
     */
    twoHandedWeapons: boolean;
    /**
     * @generated from protobuf field: bool matching_gems_only = 8
     */
    matchingGemsOnly: boolean;
    /**
     * Item IDs for favorited items.
     *
     * @generated from protobuf field: repeated int32 favorite_items = 11
     */
    favoriteItems: number[];
    /**
     * Item IDs for favorited gems.
     *
     * @generated from protobuf field: repeated int32 favorite_gems = 12
     */
    favoriteGems: number[];
    /**
     * Uniquely-identifying strings for favorited enchants.
     *
     * @generated from protobuf field: repeated string favorite_enchants = 13
     */
    favoriteEnchants: string[];
}
/**
 * @generated from protobuf message proto.SimSettings
 */
export interface SimSettings {
    /**
     * @generated from protobuf field: int32 iterations = 1
     */
    iterations: number;
    /**
     * @generated from protobuf field: int32 phase = 2
     */
    phase: number;
    /**
     * @generated from protobuf field: int64 fixed_rng_seed = 3
     */
    fixedRngSeed: bigint;
    /**
     * @generated from protobuf field: bool show_damage_metrics = 8
     */
    showDamageMetrics: boolean;
    /**
     * @generated from protobuf field: bool show_threat_metrics = 4
     */
    showThreatMetrics: boolean;
    /**
     * @generated from protobuf field: bool show_healing_metrics = 7
     */
    showHealingMetrics: boolean;
    /**
     * @generated from protobuf field: bool show_experimental = 5
     */
    showExperimental: boolean;
    /**
     * @generated from protobuf field: bool show_ep_values = 11
     */
    showEpValues: boolean;
    /**
     * @generated from protobuf field: string language = 9
     */
    language: string;
    /**
     * @generated from protobuf field: proto.Faction faction = 6
     */
    faction: Faction;
    /**
     * @generated from protobuf field: proto.DatabaseFilters filters = 10
     */
    filters?: DatabaseFilters;
}
/**
 * Contains all information that is imported/exported from an individual sim.
 *
 * @generated from protobuf message proto.IndividualSimSettings
 */
export interface IndividualSimSettings {
    /**
     * @generated from protobuf field: proto.SimSettings settings = 5
     */
    settings?: SimSettings;
    /**
     * @generated from protobuf field: proto.RaidBuffs raid_buffs = 1
     */
    raidBuffs?: RaidBuffs;
    /**
     * @generated from protobuf field: proto.Debuffs debuffs = 8
     */
    debuffs?: Debuffs;
    /**
     * @generated from protobuf field: repeated proto.UnitReference tanks = 7
     */
    tanks: UnitReference[];
    /**
     * @generated from protobuf field: proto.PartyBuffs party_buffs = 2
     */
    partyBuffs?: PartyBuffs;
    /**
     * @generated from protobuf field: proto.Player player = 3
     */
    player?: Player;
    /**
     * @generated from protobuf field: proto.Encounter encounter = 4
     */
    encounter?: Encounter;
    /**
     * @generated from protobuf field: int32 target_dummies = 9
     */
    targetDummies: number;
    /**
     * @generated from protobuf field: proto.UnitStats ep_weights_stats = 10
     */
    epWeightsStats?: UnitStats;
    /**
     * @generated from protobuf field: repeated double ep_ratios = 11
     */
    epRatios: number[];
    /**
     * @generated from protobuf field: proto.Stat dps_ref_stat = 12
     */
    dpsRefStat: Stat;
    /**
     * @generated from protobuf field: proto.Stat heal_ref_stat = 13
     */
    healRefStat: Stat;
    /**
     * @generated from protobuf field: proto.Stat tank_ref_stat = 14
     */
    tankRefStat: Stat;
}
// ...
export interface SavedTalents {
    /**
     * @generated from protobuf field: string talents_string = 1
     */
    talentsString: string;
    /**
     * @generated from protobuf field: proto.Glyphs glyphs = 2
     */
    glyphs?: Glyphs;
}
// ...
export const DatabaseFilters = new DatabaseFilters$Type();
// ...
export const SimSettings = new SimSettings$Type();
// ...
export const IndividualSimSettings = new IndividualSimSettings$Type();
// ...
class SavedTalents$Type extends MessageType<SavedTalents> {
    constructor() {
        super("proto.SavedTalents", [
            { no: 1, name: "talents_string", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "glyphs", kind: "message", T: () => Glyphs }
        ]);
    }
    create(value?: PartialMessage<SavedTalents>): SavedTalents {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.talentsString = "";
        if (value !== undefined)
            reflectionMergePartial<SavedTalents>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: SavedTalents): SavedTalents {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string talents_string */ 1:
                    message.talentsString = reader.string();
                    break;
                case /* proto.Glyphs glyphs */ 2:
                    message.glyphs = Glyphs.internalBinaryRead(reader, reader.uint32(), options, message.glyphs);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: SavedTalents, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string talents_string = 1; */
        if (message.talentsString !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.talentsString);
        /* proto.Glyphs glyphs = 2; */
        if (message.glyphs)
            Glyphs.internalBinaryWrite(message.glyphs, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message proto.SavedTalents
 */
export const SavedTalents = new SavedTalents$Type();